+++ core/tauri/src/scope/mod.rs
@@ -8,13 +8,14 @@ mod http;
 mod shell;
 
 pub use self::http::Scope as HttpScope;
-pub use fs::Scope as FsScope;
+pub use fs::{Event as FsScopeEvent, Pattern as GlobPattern, Scope as FsScope};
 #[cfg(shell_scope)]
 pub use shell::{
   ExecuteArgs, Scope as ShellScope, ScopeAllowedArg as ShellScopeAllowedArg,
   ScopeAllowedCommand as ShellScopeAllowedCommand, ScopeConfig as ShellScopeConfig,
   ScopeError as ShellScopeError,
 };
+use std::path::Path;
 
 pub(crate) struct Scopes {
   pub fs: FsScope,
@@ -24,4 +25,22 @@ pub(crate) struct Scopes {
   pub http: HttpScope,
   #[cfg(shell_scope)]
   pub shell: ShellScope,
+}
+
+impl Scopes {
+  #[allow(dead_code)]
+  pub(crate) fn allow_directory(&self, path: &Path, recursive: bool) -> crate::Result<()> {
+    self.fs.allow_directory(path, recursive)?;
+    #[cfg(protocol_asset)]
+    self.asset_protocol.allow_directory(path, recursive)?;
+    Ok(())
+  }
+
+  #[allow(dead_code)]
+  pub(crate) fn allow_file(&self, path: &Path) -> crate::Result<()> {
+    self.fs.allow_file(path)?;
+    #[cfg(protocol_asset)]
+    self.asset_protocol.allow_file(path)?;
+    Ok(())
+  }
 }