+++ examples/api/src-tauri/src/main.rs
@@ -16,8 +16,8 @@ use std::sync::atomic::{AtomicBool, Ordering};
 
 use serde::{Deserialize, Serialize};
 use tauri::{
-  api::dialog::ask, http::ResponseBuilder, CustomMenuItem, GlobalShortcutManager, Manager,
-  RunEvent, SystemTray, SystemTrayEvent, SystemTrayMenu, WindowBuilder, WindowUrl,
+  api::dialog::ask, http::ResponseBuilder, window::WindowBuilder, CustomMenuItem,
+  GlobalShortcutManager, Manager, RunEvent, SystemTray, SystemTrayEvent, SystemTrayMenu, WindowUrl,
 };
 
 #[derive(Clone, Serialize)]
@@ -129,14 +129,9 @@ fn main() {
             item_handle.set_title(new_title).unwrap();
           }
           "new" => {
-            app
-              .create_window(
-                "new",
-                WindowUrl::App("index.html".into()),
-                |window_builder, webview_attributes| {
-                  (window_builder.title("Tauri"), webview_attributes)
-                },
-              )
+            WindowBuilder::new(app, "new", WindowUrl::App("index.html".into()))
+              .title("Tauri")
+              .build()
               .unwrap();
           }
           #[cfg(target_os = "macos")]
@@ -145,7 +140,7 @@ fn main() {
 
             app
               .tray_handle()
-              .set_icon(tauri::Icon::Raw(
+              .set_icon(tauri::TrayIcon::Raw(
                 include_bytes!("../../../.icons/tray_icon_with_transparency.png").to_vec(),
               ))
               .unwrap();
@@ -156,36 +151,36 @@ fn main() {
 
             app
               .tray_handle()
-              .set_icon(tauri::Icon::Raw(
-                include_bytes!("../../../.icons/tray_icon.png").to_vec(),
+              .set_icon(tauri::TrayIcon::Raw(
+                include_bytes!("../../../.icons/tray_icon_with.png").to_vec(),
               ))
               .unwrap();
           }
           #[cfg(target_os = "linux")]
           "icon_1" => app
             .tray_handle()
-            .set_icon(tauri::Icon::File(PathBuf::from(
+            .set_icon(tauri::TrayIcon::File(PathBuf::from(
               "../../../.icons/tray_icon_with_transparency.png",
             )))
             .unwrap(),
           #[cfg(target_os = "linux")]
           "icon_2" => app
             .tray_handle()
-            .set_icon(tauri::Icon::File(PathBuf::from(
+            .set_icon(tauri::TrayIcon::File(PathBuf::from(
               "../../../.icons/tray_icon.png",
             )))
             .unwrap(),
           #[cfg(target_os = "windows")]
           "icon_1" => app
             .tray_handle()
-            .set_icon(tauri::Icon::Raw(
+            .set_icon(tauri::TrayIcon::Raw(
               include_bytes!("../../../.icons/tray_icon_with_transparency.ico").to_vec(),
             ))
             .unwrap(),
           #[cfg(target_os = "windows")]
           "icon_2" => app
             .tray_handle()
-            .set_icon(tauri::Icon::Raw(
+            .set_icon(tauri::TrayIcon::Raw(
               include_bytes!("../../../.icons/icon.ico").to_vec(),
             ))
             .unwrap(),