+++ tooling/api/src/shell.ts
@@ -29,20 +29,20 @@
  *
  * This API has a scope configuration that forces you to restrict the programs and arguments that can be used.
  *
- * ### Restricting access to the [[`open`]] API
+ * ### Restricting access to the [[open | `open`]] API
  *
  * On the allowlist, `open: true` means that the [[open]] API can be used with any URL,
  * as the argument is validated with the `^https?://` regex.
  * You can change that regex by changing the boolean value to a string, e.g. `open: ^https://github.com/`.
  *
- * ### Restricting access to the [[`Command`]] APIs
+ * ### Restricting access to the [[Command | `Command`]] APIs
  *
  * The `shell` allowlist object has a `scope` field that defines an array of CLIs that can be used.
- * Each CLI is a configuration object `{ name: string, command: string, sidecar?: bool, args?: boolean | Arg[] }`.
+ * Each CLI is a configuration object `{ name: string, cmd: string, sidecar?: bool, args?: boolean | Arg[] }`.
  *
  * - `name`: the unique identifier of the command, passed to the [[Command.constructor | Command constructor]].
  * If it's a sidecar, this must be the value defined on `tauri.conf.json > tauri > bundle > externalBin`.
- * - `command`: the program that is executed on this configuration. If it's a sidecar, it must be the same as `name`.
+ * - `cmd`: the program that is executed on this configuration. If it's a sidecar, this value is ignored.
  * - `sidecar`: whether the object configures a sidecar or a system program.
  * - `args`: the arguments that can be passed to the program. By default no arguments are allowed.
  *   - `true` means that any argument list is allowed.
@@ -59,7 +59,7 @@
  * {
  *   "scope": {
  *     "name": "run-git-commit",
- *     "command": "git",
+ *     "cmd": "git",
  *     "args": ["commit", "-m", { "validator": "\\S+" }]
  *   }
  * }
@@ -394,7 +394,7 @@ type CommandEvent =
  * or the one specified with `openWith`.
  *
  * The `openWith` value must be one of `firefox`, `google chrome`, `chromium` `safari`,
- * `open`, `start`, `xdg-open`, `gio`, gnome-open`, `kde-open` or `wslview`.
+ * `open`, `start`, `xdg-open`, `gio`, `gnome-open`, `kde-open` or `wslview`.
  *
  * @example
  * ```typescript