+++ tooling/bundler/src/bundle/windows/msi/wix.rs
@@ -404,13 +404,13 @@ pub fn build_wix_app_installer(
             .timestamp_url
             .as_ref()
             .map(|url| url.to_string()),
+          tsp: settings.windows().tsp,
         },
       )?;
     }
     Ok(())
   };
 
-  common::print_info("trying to sign app")?;
   try_sign(&app_exe_source)?;
 
   // ensure that `target/{release, debug}/wix` folder exists